+++ apps/container/src/pages/Performance/PerformancePage.vue
@@ -53,7 +53,7 @@ import { useTheme } from 'vuetify';
 import { CenteredLoading } from '@/components/ui/CenteredLoading';
 import { PaperCard } from '@/components/ui/PaperCard';
 import { PerformanceCard } from '@/components/ui/PerformanceCard';
-import { formatSeason } from '@/utils/season';
+import { formatSeason } from '@/helpers/utils/season';
 
 const theme = useTheme();
 const chartKey = ref(0);
