+++ apps/container/src/pages/Settings/SettingsPage.vue
@@ -142,8 +142,8 @@ import { computed, ref } from 'vue';
 import { CenteredLoading } from '@/components/ui/CenteredLoading';
 import { PaperCard } from '@/components/ui/PaperCard';
 import { useAliasInitials } from '@/composables/aliasInitials';
+import { buildGoogleAuthUrl } from '@/helpers/utils/runtimeConfig';
 import { useAuthStore } from '@/stores/auth';
-import { buildGoogleAuthUrl } from '@/utils/runtimeConfig';
 
 const authStore = useAuthStore();
